+++ www/editsite.cgi
@@ -6,26 +6,30 @@ if($status) exit 0
 can --site=$site admin
 setenv NOTADMIN $status
 if($?DELETE) then
- sql "$DB" 'UPDATE session set site=NULL WHERE site=$site'
- setenv C `sql -c "$DB" 'DELETE FROM site WHERE site=$site'`
- if("$C" == "" || "$C" == "0") then
-  setenv MSG "Cannot delete as in use"
-  goto done
- endif
- echo "Location: ${ENVCGI_SERVER}?MSG=Deleted"
- echo ""
- exit 0
+	if(! $?SURE) then
+		setenv MSG "Tick to say you are sure"
+		goto done
+	endif
+	sql "$DB" 'UPDATE session set site=NULL WHERE site=$site'
+	setenv C `sql -c "$DB" 'DELETE FROM site WHERE site=$site'`
+	if("$C" == "" || "$C" == "0") then
+		setenv MSG "Cannot delete as in use"
+		goto done
+	endif
+	echo "Location: ${ENVCGI_SERVER}?MSG=Deleted"
+	echo ""
+	exit 0
 endif
 if($?description) then
- if($site == 0) then
-  setenv site `sql -i "$DB" 'INSERT INTO site SET organisation=$SESSION_ORGANISATION,site=0'`
-  sql "$DB" 'INSERT INTO area SET site=$site,area="A",description="Main building"'
-  sql "$DB" 'UPDATE session SET organisation=$SESSION_ORGANISATION,site=$site WHERE session="$ENVCGI"'
- endif
- sqlwrite -o "$DB" site description
- echo "Location: ${ENVCGI_SERVER}"
- echo ""
- exit 0
+	if($site == 0) then
+		setenv site `sql -i "$DB" 'INSERT INTO site SET organisation=$SESSION_ORGANISATION,site=0'`
+		sql "$DB" 'INSERT INTO area SET site=$site,area="A",description="Main building"'
+		sql "$DB" 'UPDATE session SET organisation=$SESSION_ORGANISATION,site=$site WHERE session="$ENVCGI"'
+	endif
+	sqlwrite -o "$DB" site description
+	echo "Location: ${ENVCGI_SERVER}"
+	echo ""
+	exit 0
 endif
 done:
 echo "Content-Type: text/html"
@@ -39,7 +43,7 @@ xmlsql -d "$DB" head.html - foot.html << 'END'
 </table>
 </sql>
 <input type=submit value="Update">
-<IF NOT site=0 USER_ADMIN><input type=submit value="Delete" name=DELETE></IF>
+<IF NOT site=0 USER_ADMIN><input type=submit value="Delete" name=DELETE><input type=checkbox name=SURE title='Tick this to say you are sure'></IF>
 </form>
 </sql>
 'END'
